+++ CHANGELOG.md
@@ -5,19 +5,29 @@ All Notable changes to `Basic Value Objects` will be documented in this file
 ## NEXT - YYYY-MM-DD
 
 ### Added
-- `Float` value object
+- Nothing
 
 ### Deprecated
 - Nothing
 
 ### Fixed
-- `Boolean::TRUE()` and `Boolean::FALSE()` are now `Boolean::true()` and `Boolean::false()`.
+- Nothing
 
 ### Removed
 - Nothing
 
 ### Security
 - Nothing
+
+
+## 0.2.0 - 2015-02-14
+
+### Added
+- `Float` value object
+
+### Fixed
+- `Boolean::TRUE()` and `Boolean::FALSE()` are now `Boolean::true()` and `Boolean::false()`
+
 
 ## 0.1.0 - 2015-02-10